+++ src/DuetAPI/Machine/Job/Layer.cs
@@ -11,22 +11,22 @@ namespace DuetAPI.Machine
         /// <summary>
         /// Duration of the layer (in s or null if unknown)
         /// </summary>
-        public double? Duration { get; set; }
+        public float? Duration { get; set; }
 
         /// <summary>
         /// Height of the layer (in mm or null if unknown)
         /// </summary>
-        public double? Height { get; set; }
+        public float? Height { get; set; }
 
         /// <summary>
         /// Actual amount of filament extruded during this layer (in mm)
         /// </summary>
-        public double[] Filament { get; set; } = new double[0];
+        public float[] Filament { get; set; } = new float[0];
 
         /// <summary>
         /// Fraction of the file printed during this layer (0..1 or null if unknown)
         /// </summary>
-        public double? FractionPrinted { get; set; }
+        public float? FractionPrinted { get; set; }
 
         /// <summary>
         /// Creates a clone of this instance
@@ -38,7 +38,7 @@ namespace DuetAPI.Machine
             {
                 Duration = Duration,
                 Height = Height,
-                Filament = (double[])Filament.Clone(),
+                Filament = (float[])Filament.Clone(),
                 FractionPrinted = FractionPrinted
             };
         }
