+++ src/DuetAPI/Machine/Move/Idle.cs
@@ -10,12 +10,12 @@ namespace DuetAPI.Machine
         /// <summary>
         /// Idle timeout after which the stepper motor currents are reduced (in s)
         /// </summary>
-        public double Timeout { get; set; } = 30;
+        public float Timeout { get; set; } = 30F;
         
         /// <summary>
         /// Motor current reduction factor (on a scale between 0 to 1)
         /// </summary>
-        public double Factor { get; set; } = 0.3;
+        public float Factor { get; set; } = 0.3F;
 
         /// <summary>
         /// Creates a clone of this instance
